Move utility classes into own package
