+++ bugshaker/src/main/java/com/github/stkent/bugshaker/EnvironmentCapabilitiesProvider.java
@@ -23,6 +23,8 @@ import android.net.Uri;
 import android.support.annotation.NonNull;
 import android.text.TextUtils;
 
+import com.github.stkent.bugshaker.utilities.Logger;
+
 import java.util.ArrayList;
 import java.util.List;
