+++ bugshaker/src/main/java/com/github/stkent/bugshaker/screenshots/MapScreenshotProvider.java
@@ -27,7 +27,7 @@ import android.support.annotation.NonNull;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.github.stkent.bugshaker.ActivityUtils;
+import com.github.stkent.bugshaker.utilities.ActivityUtils;
 import com.google.android.gms.maps.MapView;
 
 import java.util.ArrayList;
